# Conflicts:
#	pyemma/coordinates/transform/pca.py
#	pyemma/coordinates/transform/tica.py

+++ pyemma/msm/estimation/sparse/mle_trev.pyx
@@ -13,7 +13,7 @@ import warnings
 import pyemma.util.exceptions
 
 cdef extern from "_mle_trev.h":
-  int _mle_trev_sparse(double * const T_data, const long long * const CCt_data, const long long * const i_indices, const long long * const j_indices, const int len_CCt, const long long * const sum_C, const int dim, const double maxerr, const int maxiter)
+  int _mle_trev_sparse(double * const T_data, const double * const CCt_data, const int * const i_indices, const int * const j_indices, const int len_CCt, const double * const sum_C, const int dim, const double maxerr, const int maxiter)
 
 def mle_trev(C, double maxerr = 1.0E-12, int maxiter = 1000000):
 
@@ -23,26 +23,26 @@ def mle_trev(C, double maxerr = 1.0E-12, int maxiter = 1000000):
   assert pyemma.msm.estimation.is_connected(C, directed=True), 'C must be strongly connected'
   
   C_sum_py = C.sum(axis=1).A1
-  cdef numpy.ndarray[long long, ndim=1, mode="c"] C_sum = C_sum_py.astype(numpy.int64, order='C', copy=False)
+  cdef numpy.ndarray[double, ndim=1, mode="c"] C_sum = C_sum_py.astype(numpy.float64, order='C', copy=False)
 
   CCt = C+C.T
   # convert CCt to coo format 
   CCt_coo = CCt.tocoo()
   n_data = CCt_coo.nnz
-  cdef numpy.ndarray[long long, ndim=1, mode="c"] CCt_data =  CCt_coo.data.astype(numpy.int64, order='C', copy=False)
-  cdef numpy.ndarray[long long, ndim=1, mode="c"] i_indices = CCt_coo.row.astype(numpy.int64, order='C', copy=True)
-  cdef numpy.ndarray[long long, ndim=1, mode="c"] j_indices = CCt_coo.col.astype(numpy.int64, order='C', copy=True)
+  cdef numpy.ndarray[double, ndim=1, mode="c"] CCt_data =  CCt_coo.data.astype(numpy.float64, order='C', copy=False)
+  cdef numpy.ndarray[int, ndim=1, mode="c"] i_indices = CCt_coo.row.astype(numpy.intc, order='C', copy=True)
+  cdef numpy.ndarray[int, ndim=1, mode="c"] j_indices = CCt_coo.col.astype(numpy.intc, order='C', copy=True)
   
   # prepare data array of T in coo format
-  cdef numpy.ndarray[double, ndim=1, mode="c"] T_data = numpy.zeros(n_data, dtype=numpy.double, order='C')
+  cdef numpy.ndarray[double, ndim=1, mode="c"] T_data = numpy.zeros(n_data, dtype=numpy.float64, order='C')
   
   err = _mle_trev_sparse(
         <double*> numpy.PyArray_DATA(T_data),
-        <long long*> numpy.PyArray_DATA(CCt_data),
-        <long long*> numpy.PyArray_DATA(i_indices),
-        <long long*> numpy.PyArray_DATA(j_indices),
+        <double*> numpy.PyArray_DATA(CCt_data),
+        <int*> numpy.PyArray_DATA(i_indices),
+        <int*> numpy.PyArray_DATA(j_indices),
         n_data,
-        <long long*> numpy.PyArray_DATA(C_sum),
+        <double*> numpy.PyArray_DATA(C_sum),
         CCt.shape[0],
         maxerr,
         maxiter)
